+++ app/Livewire/Focal/Implementations/ViewBeneficiary.php
@@ -658,8 +658,10 @@ class ViewBeneficiary extends Component
             if ($identity) {
                 if ($this->image_file_path) {
                     $file = $this->image_file_path->store(path: 'credentials');
-                    if (Storage::exists($identity->image_file_path)) {
-                        Storage::delete($identity->image_file_path);
+                    if ($identity->image_file_path) {
+                        if (Storage::exists($identity->image_file_path)) {
+                            Storage::delete($identity->image_file_path);
+                        }
                     }
                     $identity->fill([
                         'image_file_path' => $file,
@@ -675,8 +677,10 @@ class ViewBeneficiary extends Component
             if ($special_case) {
                 if ($this->reason_image_file_path) {
                     $file = $this->reason_image_file_path->store(path: 'credentials');
-                    if (Storage::exists($special_case->image_file_path)) {
-                        Storage::delete($special_case->image_file_path);
+                    if ($special_case->image_file_path) {
+                        if (Storage::exists($special_case->image_file_path)) {
+                            Storage::delete($special_case->image_file_path);
+                        }
                     }
                     $special_case->fill([
                         'image_description' => $this->image_description,
